% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeArgument}
\alias{makeArgument}
\title{Make argument object.}
\usage{
makeArgument(name, shortcut = NULL, alias = NULL, is.flag = FALSE,
  default = NULL, type = NULL, required = FALSE, help)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Name of the argument.}

\item{shortcut}{[\code{character(1)}]\cr
Shortcut name of the argument. Default is \code{NULL}, i. e., no shortcut.}

\item{alias}{[\code{character}]\cr
Character vector of alias names for \code{name} argument. Default is \code{NULL},
i. e., no aliases at all.}

\item{is.flag}{[\code{logical(1)}]\cr
Set this to \code{TRUE}, if the argument is a flag, i. e., it has no value.}

\item{default}{[\code{any}]\cr
Default value for the argument. Keep in mind, that this must be of the corresponding
\code{type}. If no default value is defined, which is the default, the user has
to provide a value.}

\item{type}{[\code{character(1)}]\cr
The type of the argument. Argument values are coerced to the type. Keep in mind, that
this argument is neglected if \code{is.flag} is set to \code{TRUE}.}

\item{required}{[\code{logical(1)}]\cr
Set this to \code{TRUE}, if the user must provide this argument. Default ist \code{FALSE}.}

\item{help}{[\code{character(1)}]\cr
Helping text.}
}
\value{
[\code{RargsArgument}] Argument object.
}
\description{
Make argument object.
}

