% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseArguments.R
\name{parseArguments}
\alias{parseArguments}
\title{Method to parse arguments.}
\usage{
parseArguments(arg.set, raw.args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{arg.set}{[\code{RargsArgumentSet}]\cr
Set of arguments.}

\item{raw.args}{[\code{character}]\cr
Character vector of raw command arguments, i. e., arguments passed by the
user to the program. Default ist \code{\link[base]{commandArgs}(trailingOnly = TRUE)}.}
}
\value{
[\code{RargsParsedArguments}]
  List of parsed arguments.
}
\description{
Method to parse arguments.
}

