% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeArgumentSet}
\alias{makeArgumentSet}
\title{Make set of arguments.}
\usage{
makeArgumentSet(...)
}
\arguments{
\item{...}{[\code{RargsArgument}]\cr
Arguments.}
}
\value{
[\code{RargsArgumentSet}]
   Set of arguments ready to pass to \code{\link{parseArguments}}.
}
\description{
Make set of arguments.
}

